Add synchronization to platform target project handler
The PptpTargetProjectHandler was called simultanously from several
threads. It was not really designed to cope with that (files are
checked for existence and then created if missing) which resulted
in sporadic exceptions (file already exist). The remedy was to
synchronize the public methods. It's reliable since the handler
is a singleton.