Ignore MacOS specific "extended attributes" file during unpack
MacOS tar adds a special file with a name that starts with
"./._" to the root of a tar file. This caused problems when
unpacking modules since a module is supposed to contain one singe
directory in the root of the tar archive.

This commit ensures that the MacOS specific file is ignored.