Use a more recent version of jackson from com.fasterxml
The Jackson framwork for XML/JSON serialization has moved from
org.codehaus.jackson to com.fasterxml.jackson. This commit contains
the changes needed to use the new namespace and also updates to
from version 1.6.0 to 2.3.3.

The new version is not present in the Orbit repository so its now
instead fetched from maven central.