Show more than one version and show supported flag
The current dialog where a module can be selected from the forge
is severely limited in that it only displays the latest version. This
often means that the version currently supported by Puppet Labs is
invisible.

This commit changes so that the supported version is noted as such
and is displayed at the very top followed by the rest, listed in
ascending order by name and descending order by version.