[GEP-248] Import from github doesn't automatically "share" the project
This commit doesn't really fix the problem per se, it has been fixed
in egit, but it ensures that the user doesn't get silly questions
about overwriting the content that was just imported. A minor glitch
that caused a dialog popup asking the user if the xtext nature should
be added was also fixed here.