(GEP-157) Replace hardcoded attribute cluster size with preference
This commit removes the constant ATTRIBUTE_OPERATIONS_CLUSTER_SIZE
with a value of 20. The value is instead picked from the
IBreakAndAlignAdvice.clusterSize() which corresponds to the
preference setting of:

PP -> Formatter -> Break and Align -> Max Alignment Padding

This preference defines the max when aligning definiton arguments
and assignments. I see no reason why it shouldn't also apply to
attributes.