(GEP-551) Warn about deprecated variable names
A Name starting with an upper case letter or a digit (but is't
a decimal number without leading zeroes) will become deprecated
in future versions of puppet. This commit ensures that Geppetto
can warn about such names (optional setting but the default is
to warn) and that such names will result in error when the future
platform is used.