Ensure that a missing puppet-lint gem yields an exception
The puppet-lint gem will return non-zero exit codes when errors
are found during a lint run. A non-zero exit code is also returned
when there puppet-lint gem cannot be found.

This commit ensures that a missing puppet-lint gem results in
an exception. It was previously ingored since no errors or warnings
were parsed from the output.