(GEP-185) Gepetto 4.3 does not warn about incompatible JVM
MacOS/X refuses to create a display unless it's the main thread that
executes. This is regardless of -XstartInFirstThread or use of
Display.asyncExec().

This commit changes the JVM version check so that it instead uses
the IStartup extension point which gets called after the Workbench
has been fully initialized and a Display has been created.