(GEP-187) Geppetto fails with latest version of puppet-lint gem
The Geppetto puppet-lint extension was using a set of hard-coded
puppet-lint checks to enable inverting the set of checks to be
executed. This is no longer necessary since puppet-lint now have
a --only-checks option. The hard-coded set also led to problems
when some options were removed in puppet-lint 1.0.0.

This commit removes the hard-coded checks completely and instead
leverages the --only-checks option. Parameters can be validated
using a CHECK_PATTERN regular expression.