(GEP-192) Fix failing test
A test relied on that something found by the scanner was reported
as "Not found on search path" although the declaration was found
in the same file as the reference. This commit adds a file that
makes a reference from another file, thus testing both cases
(found in search path and not found).