(GEP-212) Create a Geppetto uber-jar with shaded entries
This commit adds a new Maven module that uses the maven
'shade' plugin to avoid conflicts in com.guice.inject and
org.antlr when using Geppetto in a Jenkins instance. This
jar is needed since the 'maskClasses' functionality provided
by the hpi plug-in seems to be ignored when running on
Jenkins slaves.