(GEP-214) Validation produces errors on functional modules
The report of dependencies that are defined more than once in the
metadata.json was not configurable. Same thing for duplicate
parameters in manifest definitions. This commit adds the needed
preferences with a default setting of WARNING. 

This commit also changes the metadata.json parser so that it allows
null in place of string literals.